Change tile set values to 8 bits

+++ Editor/include/TileMap/TileSet.h
@@ -2,7 +2,6 @@
 
 #include "Core.h"
 
-#include <string>
 #include <vector>
 #include <array>
 
@@ -20,16 +19,16 @@ public:
 public:
 	TileSet() = default;
 
-	std::array<glm::vec2, 4> GetTexCoords(const uint32_t tileId);
-	const Terrain& GetTerrain(const uint32_t tileId);
+	std::array<glm::vec2, 4> GetTexCoords(const uint8_t tileId);
+	const Terrain& GetTerrain(const uint8_t tileId);
 
 public:
 	Ref<GLTexture> texture;
 	std::vector<Terrain> terrains;
 
-	uint32_t firstGid = 1;
-	uint32_t tileWidth = 0;
-	uint32_t tileHeight = 0;
-	uint32_t tileCount = 0;
-	uint32_t columnCount = 0;
+	uint8_t firstGid = 1;
+	uint8_t tileWidth = 0;
+	uint8_t tileHeight = 0;
+	uint8_t tileCount = 0;
+	uint8_t columnCount = 0;
 };
